requirements.txt
requests
git+https://github.com/Rapptz/discord.py
python-dotenv 
quart
aiofiles
psutil
asyncio
colorama
jishaku
pillow
python-dateutil
aiosqlite
discord-ext-ipcx
quartcord
quart_cors